requirements.txt
# ===============================
# STREAMLIT UI
# ===============================
streamlit==1.32.2

# ===============================
# INSTAGRAM AUTOMATION
# ===============================
instagrapi==1.19.6
requests==2.31.0
urllib3==2.1.0

# ===============================
# IMAGE & VIDEO PROCESSING
# ===============================
Pillow==10.2.0
moviepy==1.0.3
numpy==1.26.4
opencv-python==4.9.0.80
imageio


# ===============================
# VIDEO BACKEND (FFMPEG BRIDGE)
# ===============================
imageio==2.34.0
imageio-ffmpeg==0.4.9

# ===============================
# AUDIO (OPTIONAL – SAFE)
# ===============================
pydub==0.25.1

# ===============================
# DATE / TIME / SCHEDULING
# ===============================
python-dateutil==2.9.0.post0
pytz==2024.1

# ===============================
# UTILITIES
# ===============================
tqdm==4.66.1
python-dotenv==1.0.1

# ===============================
# REEL / LINK DOWNLOAD
# ===============================
yt-dlp==2024.03.10

# ===============================
# AI / CAPTION GENERATION
# ===============================
openai==1.10.0
groq==0.11.0 
httpx==0.27.0

# ===============================
# COMPATIBILITY
# ===============================
typing-extensions==4.9.0
